css/css-fonts/parsing/font-feature-settings-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-feature-settings with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-feature-settings-prop">
<meta name="assert" content="font-feature-settings supports only the grammar 'normal | <feature-tag-value> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("font-feature-settings", "auto");
test_invalid_value("font-feature-settings", "normal 'tag'");
test_invalid_value("font-feature-settings", "123");
test_invalid_value("font-feature-settings", "'tag' on off");
</script>
</body>
</html>
